some more improvements for numCombinations()

+++ src/main/kotlin/io/craigmiller160/kotlin/commons/math/MathFunctions.kt
@@ -1,5 +1,7 @@
 package io.craigmiller160.kotlin.commons.math
 
+import java.math.BigInteger
+
 fun factorial(value: Long): Long{
     return factorialRecursive(value, value)
 }
@@ -9,6 +11,6 @@ private fun factorialRecursive(factor: Long, value: Long): Long{
     return if(newValue > 0) factorialRecursive(factor * newValue, newValue) else factor
 }
 
-fun numCombinations(numRepetition: Int, numChoices: Int): Int{
-    return Math.pow(numChoices.toDouble(), numRepetition.toDouble()).toInt()
+fun numCombinations(numRepetition: Int, numChoices: Int): Long{
+    return Math.pow(numChoices.toDouble(), numRepetition.toDouble()).toLong()
 }
